Fixed incorrect mouse cursor scaling, which breaks hover/tooltips/inventory
The mouse cursor coordinate conversion was using “getScreenWidth()” rather than “getWidth()” (i.e. using the screen size, rather than the window size), which was resulting in an incorrect scaling factor. This fixes it, in all the places it’s used.

This should fix:

#440 - Inventory Items are far off screen to the right when selected
#425 - Inventory Movement Issues
#336 - Items and cursor turn invisible when trying to re-order items in inventory
#214 - MacOS Distortion [Unplayable] on ALL Mod Versions
#37 - Bug: Mouse icon moves to other Minecraft screen, or vanishes when in inventory

This issue is a show-stopper on macOS, and possibly on Windows with a Hi-DPI display, so I believe it’s quite high priority.